apps/api/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
openai==1.12.0

# Database
motor==3.3.2  # Async MongoDB
qdrant-client==1.7.3

# Vector & Embeddings
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Logging & Monitoring
structlog==24.1.0

# Utils
PyYAML==6.0.1
aiofiles==23.2.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
ruff==0.2.0
mypy==1.8.0
